Added time data output to file
Changed output from stdout to file named by the signal which produced output.
The signal state sends a time stamp to the appropriate file, therefore making a
way to track machine cycle counts between different machines. The idea is that
another program/script will look at the files and count -- for the cycle
counters -- the occurrence of ON states with the first column matching the
machine name and the timestamps matching the date range entered in the query.
For hour meters, it will find all rows matching the machine name, then
accumulate the difference of OFF and ON states within the matching time stamps
in the date range entered in the query.

Signed-off-by: Ian Graeme Hilt <ian.hilt@gmail.com>