Introduced functions to handle output
output_on_time() and output_off_time() handle the output from the while loop to
the file. I think this will help condense the body of the program and improve
the logical flow of the code.

I collected all of the signal names into const char's to preclude the need to
enlarge the while loop and/or the output functions. As is stated in the comment
for these, there may be a better way to do this; not sure yet.

Signed-off-by: Ian Graeme Hilt <ian.hilt@gmail.com>